ptuxrepo: set setgid bit on initialized repositories
Initialize repositories with the setgit bit set on the top
directory so that files and directories created within inherit
the group ownership of the repository.

An easy way to create a repository owned by a group is to create
the top directory manually and setting its group, before calling
ptuxrepo init. E.g.:

    $ mkdir /srv/repo
    $ chgrp devgroup /srv/repo
    $ ptuxrepo init /srv/repo