ptuxversion: rewrite with sophistication and potential
The previous version was getting unwieldy; its function obscured
by all the cleverness it takes to write something in shell
script. Introduce a more straightforwardly-written Python
version, with a test suite. It's at least as featureful as
previous version, but with better respect for the authority of
tags to override calculated versions, and better demangling of
tag names according to DEP14.