ptuxchange: use rev-parse to check for git
Checking only for a .git directory in the working directory means
that we have to run ptuxbuild from the topmost directory of the
repository. Not all projects are structured that way.

If the check for .git fails, we try git rev-parse --git-dir
before we declare that we're not in a git-controlled directory.
This lets us confirm from any subdirectory of the project.

Signed-off-by: Bill Gatliff <bgat@billgatliff.com>